Function draw2d::camera::default_camera_controls

pub fn default_camera_controls(
    camera: &mut OrthoCamera, 
    event: &WindowEvent
) -> bool

A really simple input handler for a camera which just works for a demo.

Real applications will almost certainly prefer a more refined camera controller which provides a more smooth experience.

Returns true when the camera has been changed in some way. Often this is used to trigger a matrix update for the graphics subsystem.