Module draw2d::graphics::draw2d

Re-exports

pub use self::descriptor_sets::UniformBufferObject;

pub use self::layer::Layer;

pub use self::layer::LayerHandle;

pub use self::layer::StackedLayers;

pub use self::texture_atlas::TextureHandle;

pub use self::vertex::Vertex;

Modules

commands

descriptor_sets

graphics_pipeline

layer
A layer is a collection of vertices which are all rasterized with the same texture.

texture_atlas
This module defines structures and functions for managing a 'global' texture atlas which can be bound once per frame regardless of how many different textures are used by the application.

vertex

Structs

Draw2d
Resources used to render triangles

Type Definitions

Mat4