Struct draw2d::graphics::frame::Frame

pub struct Frame {
    pub sync: FrameSync,
    pub descriptor: FrameDescriptor,
    pub vertex_buffer: CpuBuffer,
    pub command_pool: ReusableCommandPool,
    pub framebuffer: Framebuffer,
    // some fields omitted
}

All per-frame resources and synchronization for this application.

Fields
sync: FrameSyncdescriptor: FrameDescriptorvertex_buffer: CpuBuffercommand_pool: ReusableCommandPoolframebuffer: Framebuffer
Implementations

impl Frame

pub fn create_n_frames(
device: &Arc<Device>,
framebuffers: &[Framebuffer]
) -> Result<Vec<Option<Self>>>

Create a collection of frames with resource debug names based on the frame index.

pub fn new<Name>(
device: Arc<Device>,
framebuffer: Framebuffer,
name: Name
) -> Result<Self> where
Name: Into<String> + Clone,

Create a new frame.

Frames do not own framebuffers, it is the responsibility of the application to ensure no Frame instances are used after the swapchain has been dropped.

pub fn begin_frame(&mut self) -> Result<()>

Begin the frame’s rendering operations.

Blocks until the previous render with this frame has finished. Resets the command pool used by this frame.

pub fn submit_graphics_commands(&mut self, command_buffers: &[CommandBuffer])

Submit command buffers to be added to the graphics queue when the frame is finished by the frame context.

pub fn finish_frame(&mut self, image_available: Semaphore) -> Result<Semaphore>

Finish the frame by submitting all command buffers to the graphics queue and a semaphore which signals when rendering to the framebuffer is complete.