Struct draw2d::graphics::layer::LayerStack

pub struct LayerStack { /* fields omitted */ }

A collection of ordered layers for rendering.

Implementations

impl LayerStack

pub fn new() -> Self

Create a new stack with zero visible layers.

pub fn add_layer_to_top(&mut self) -> LayerHandle

Add a layer to be rendered on top of all existing layers.

pub fn add_layer_to_bottom(&mut self) -> LayerHandle

Add a layer to be rendered under all existing layers

pub fn layers(&self) -> Vec<&Layer>

Return the set of all layer references in their render order.

pub fn get_layer_mut(&mut self, handle: &LayerHandle) -> Option<&mut Layer>

Get the layer referenced by a handle.

Returns None if the handle is invalid.

pub fn vertices(&self) -> Vec<&[Vertex2d]>

Get the slice of all vertices for all layers and batches in order.

This can be used to build a vertex buffer when rendering.

The layout of the vertices is like the following:

Layer

Batch vertices

Batch vertices

Layer

Batch vertices

Layer

Batch vertices

Batch vertices

Batch vertices