Struct draw2d::graphics::Graphics

pub struct Graphics {
    pub texture_atlas: GpuAtlas,
    pub clear_color: [f32; 4],
    pub device: Arc<Device>,
    // some fields omitted
}

The application’s graphics subsystem.

Fields
texture_atlas: GpuAtlas

The graphics subsystem’s texture atlas.

clear_color: [f32; 4]

the color used to clear the screen

device: Arc<Device>

The vulkan device used by all resources in the graphics subsystem.

Implementations

impl Graphics

pub fn new(window_surface: &dyn WindowSurface) -> Result<Self>

Instantiate the graphics subsystem.

pub fn add_layer_to_top(&mut self) -> LayerHandle

Add a new graphics layer to the top of the rendering stack.

This layer will be rendered above all other existing layers.

pub fn add_layer_to_bottom(&mut self) -> LayerHandle

Add a new graphics layer to the bottom of the rendering stack.

This layer will be rendered below all other existing layers.

pub fn get_layer_mut(&mut self, layer_handle: &LayerHandle) -> &mut Layer

Return a mutable reference to the layer referenced by the handle

PANICs if the layer handle doesn’t refer to an actual layer.

pub fn render(&mut self, window_surface: &dyn WindowSurface) -> Result<()>

Render a single frame to the screen.

pub fn rebuild_swapchain(
&mut self,
window_surface: &dyn WindowSurface
) -> Result<()>

Replace the swapchain and all dependent resources in the Triangle subsystem.