Module draw2d::graphics::vulkan

This module contains functions and structures for handling vulkan resources.

Re-exports

pub use self::device::Device;

pub use self::instance::Instance;

pub use self::swapchain::Swapchain;

pub use self::window_surface::WindowSurface;

Modules

buffer

command_pool
This module provides structures for managing a collection of command buffers for a given command pool.

device
This module provides functions for picking a physical device and creating the logical device.

device_allocator
This module defines traits and implementations for managing device (gpu) memory.

ffi
Convenience functions for interacting with ffi calls to vulkan.

instance
Functions for creating an instance with extensions.

shader_module
This module defines the ShadeModule abstraction which makes it easy to create vulkan shader modules directly with the rust include_bytes macro.

swapchain
This module defines the Swapchain abstraction and related Vulkan resources.

texture

window_surface
This module defines the WindowSurface trait which defines the behavior Vulkan requires from a window implementation.