Function draw2d::graphics::vulkan::buffer::copy_full_buffer

pub unsafe fn copy_full_buffer<Source, Destination>(
    device: &Device, 
    command_buffer: CommandBuffer, 
    src: &Source, 
    dst: &Destination
) -> Result<CommandBuffer> where
    Source: Buffer,
    Destination: Buffer, 

Write commands to copy the full source buffer to the destination buffer.

Unsafe because this method does not check the destination’s size. It is the responsibility of the application to ensure the destination buffer is at least as large as the source buffer.