Trait draw2d::graphics::vulkan::buffer::Buffer

pub trait Buffer {
    unsafe fn raw(&self) -> Buffer;

    unsafe fn allocation(&self) -> &Allocation;

    fn size_in_bytes(&self) -> u64;
}

Required methods

unsafe fn raw(&self) -> Buffer

The raw vulkan handle for the buffer. Should not be copied because implementations are allowed to invalidate the raw buffer value.

Unsafe because it is the responsibility of the caller to ensure that the handle will live for the duration of it’s usage. (by checking the specific implementation)

unsafe fn allocation(&self) -> &Allocation

The raw handle to the buffer’s underlying memory allocation.

fn size_in_bytes(&self) -> u64

The size of the underlying gpu memory in bytes.

Loading content...