Function draw2d::graphics::vulkan::device_allocator::build_standard_allocator

pub fn build_standard_allocator(
    ash_instance: Instance, 
    logical_device: Device, 
    physical_device: PhysicalDevice
) -> Box<impl DeviceAllocator>

Build the standard allocator implementation.

The return is a boxed impl so that consumers are not dependent on the specific implementation (often the full type is unwieldy because it is a composition of DeviceAllocator implementations).

The caller is responsible for keeping the ash instance, logical device, and physical device alive for at least as long as the allocator exists.