Module draw2d::graphics::vulkan::instance

Functions for creating an instance with extensions.

The Instance struct holds the ash entry and ash instance along with the debug callback. This is convenient because the application needs to hold references to all of this data, but it's unwieldy to have separate handles to each constantly floating around.

Example

let instance = Instance::new(
    &glfw.get_required_instance_extensions().unwrap()
)?;

Modules

extensions
Functions to check that a set of extensions are supported by the vulkan instance.

layers
Functions to check if a set of vulkan layers are available for the instance.

Structs

Instance
Hold all of the instance-related objects and drop them in the correct order.

Functions

create_debug_callback
Create the vulkan debug callback for validation.

create_instance
Create a Vulkan instance with the required extensions. Returns an Err() if any required extensions are unavailable.

debug_callback⚠

required_layers
The validation layers required by this application.