Module draw2d::graphics::vulkan::swapchain::selection

Functions for selecting correct swapchain properties for this application.

Functions

choose_image_count
Choose the number of images to use in the swapchain based on the min and max numbers of images supported by the device. querying surface capabilities is safe in this context because the physical device will not be selected unless it supports the swapchain extension

choose_present_mode
Choose a presentation mode for the swapchain based on the window and chosen physical device.

choose_surface_format
Choose a surface format for the swapchain based on the window and chosen physical device.

choose_swap_extent
Choose the swap extent for the swapchain based on the window's framebuffer size. Getting surface capabilities is safe because suppport for the swapchain extenstion is verified when picking a physical device

clamp
Clamp a value between a minimum and maximum bound.