Function draw2d::graphics::vulkan::swapchain::selection::choose_present_mode

pub fn choose_present_mode(
    window_surface: &dyn WindowSurface, 
    physical_device: &PhysicalDevice
) -> PresentModeKHR

Choose a presentation mode for the swapchain based on the window and chosen physical device.

checking presentation modes is safe because support for the swapchain extension is verified when picking a physical device